+++ frontend/src/components/FormPriorityToggleButtons.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import ToggleButton from "@mui/material/ToggleButton";
-import Tooltip from "@mui/material/Tooltip";
 import LowPriorityIcon from "@mui/icons-material/KeyboardArrowDown";
 import MediumPriorityIcon from "@mui/icons-material/Remove";
 import HighPriorityIcon from "@mui/icons-material/KeyboardArrowUp";
@@ -19,27 +18,21 @@ const FormPriorityToggleButtons = (props: FormPriorityToggleButtonsProps): JSX.E
 
   return (
     <ToggleButtonGroup value={value} exclusive onChange={onChangeHandler} size='small'>
-      <ToggleButton value={"low"} color='success'>
-        <Tooltip title={"Low Priority"} disableInteractive>
-          <LowPriorityIcon fontSize='small' />
-        </Tooltip>
+      <ToggleButton value={"low"} color='success' sx={{ textTransform: "none" }}>
+        <LowPriorityIcon fontSize='small' />
+        Low
       </ToggleButton>
-      <ToggleButton value={"medium"} color='warning'>
-        <Tooltip title={"Medium Priority"} disableInteractive>
-          <MediumPriorityIcon fontSize='small' />
-        </Tooltip>
+      <ToggleButton value={"medium"} color='warning' sx={{ textTransform: "none" }}>
+        <MediumPriorityIcon fontSize='small' />
+        Medium
       </ToggleButton>
-      <ToggleButton value={"high"} color='error'>
-        <Tooltip title={"High Priority"} disableInteractive>
-          <HighPriorityIcon fontSize='small' />
-        </Tooltip>
+      <ToggleButton value={"high"} color='error' sx={{ textTransform: "none" }}>
+        <HighPriorityIcon fontSize='small' />
+        High
       </ToggleButton>
-      <ToggleButton value={"highest"} color='error'>
-        <Tooltip title={"Highest Priority"} disableInteractive>
-          <div>
-            <HighestPriorityIcon />
-          </div>
-        </Tooltip>
+      <ToggleButton value={"highest"} color='error' sx={{ textTransform: "none" }}>
+        <HighestPriorityIcon />
+        Urgent
       </ToggleButton>
     </ToggleButtonGroup>
   );